requirements.txt
graphsurgeon==0.4.5
numpy==1.19.4
onnx==1.6.0
onnxruntime-gpu==1.5.2
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==5.1.0
tqdm==4.56.1
torch==1.8.1
torchvision==0.2.2.post3
pycuda==2020.1
tensorrt==7.1.3.0
scikit-image==0.17.2
face-alignment==1.3.3